@@ -1,6 +1,6 @@
 {
   "name": "parse-dashboard",
-  "version": "3.0.0",
+  "version": "3.2.1",
   "repository": {
     "type": "git",
     "url": "https://github.com/ParsePlatform/parse-dashboard"
@@ -35,34 +35,30 @@
     "LICENSE"
   ],
   "dependencies": {
-    "@babel/runtime": "7.15.3",
+    "@babel/runtime": "7.15.4",
     "bcryptjs": "2.3.0",
     "body-parser": "1.19.0",
-    "codemirror-graphql": "github:timsuchanek/codemirror-graphql#details-fix",
     "commander": "6.2.1",
     "connect-flash": "0.1.1",
     "cookie-session": "2.0.0-beta.3",
     "copy-to-clipboard": "3.2.0",
-    "create-react-class": "15.7.0",
     "csurf": "1.11.0",
     "express": "4.17.1",
-    "graphql": "14.6.0",
-    "graphql-playground-react": "1.7.20",
+    "graphiql": "1.4.2",
+    "graphql": "15.4.0",
     "history": "4.10.1",
     "immutable": "4.0.0-rc.9",
     "immutable-devtools": "0.1.5",
-    "inquirer": "8.1.2",
+    "inquirer": "8.1.3",
     "js-beautify": "1.14.0",
-    "json-file-plus": "3.2.0",
-    "otpauth": "7.0.5",
+    "otpauth": "7.0.6",
     "package-json": "6.5.0",
-    "parse": "3.3.0",
+    "parse": "3.3.1",
     "passport": "0.4.1",
     "passport-local": "1.0.0",
     "prismjs": "1.25.0",
     "prop-types": "15.7.2",
     "qrcode": "1.4.4",
-    "query-string": "6.14.1",
     "react": "16.14.0",
     "react-ace": "9.4.3",
     "react-dnd": "10.0.2",
@@ -73,9 +69,10 @@
     "react-popper-tooltip": "4.3.0",
     "react-redux": "5.1.2",
     "react-router": "5.2.1",
-    "react-router-dom": "5.2.1",
+    "react-router-dom": "5.3.0",
     "regenerator-runtime": "0.13.8",
-    "semver": "7.3.4"
+    "semver": "7.3.4",
+    "typescript": "4.4.3"
   },
   "devDependencies": {
     "@actions/core": "1.2.6",
@@ -85,6 +82,12 @@
     "@babel/plugin-transform-runtime": "7.8.3",
     "@babel/preset-env": "7.9.0",
     "@babel/preset-react": "7.8.3",
+    "@semantic-release/changelog": "5.0.1",
+    "@semantic-release/commit-analyzer": "8.0.1",
+    "@semantic-release/git": "9.0.0",
+    "@semantic-release/github": "7.2.3",
+    "@semantic-release/npm": "7.1.3",
+    "@semantic-release/release-notes-generator": "9.0.3",
     "all-node-versions": "8.0.0",
     "babel-eslint": "10.1.0",
     "babel-loader": "8.1.0",
@@ -100,18 +103,19 @@
     "jest": "24.8.0",
     "madge": "5.0.1",
     "marked": "0.8.2",
-    "node-sass": "5.0.0",
-    "null-loader": "3.0.0",
+    "node-sass": "6.0.0",
+    "null-loader": "4.0.1",
     "path-to-regexp": "3.2.0",
     "puppeteer": "3.0.0",
     "react-test-renderer": "16.13.1",
     "request": "2.88.2",
     "request-promise": "4.2.5",
-    "sass-loader": "10.1.1",
+    "sass-loader": "11.1.1",
+    "semantic-release": "17.4.6",
     "style-loader": "1.1.2",
     "svg-prep": "1.0.4",
-    "webpack": "4.42.1",
-    "webpack-cli": "3.3.10",
+    "webpack": "5.53.0",
+    "webpack-cli": "4.0.0",
     "yaml": "1.10.0"
   },
   "scripts": {
@@ -127,13 +131,14 @@
     "generate": "node scripts/generate.js",
     "prepare": "webpack --config webpack/publish.config.js --progress",
     "start": "node ./Parse-Dashboard/index.js",
-    "madge:circular": "node_modules/.bin/madge ./src --circular"
+    "madge:circular": "node_modules/.bin/madge ./src --circular",
+    "semantic-release": "semantic-release"
   },
   "bin": {
     "parse-dashboard": "./bin/parse-dashboard"
   },
   "engines": {
-    "node": ">=12.0.0 <16.0.0"
+    "node": ">=12.0.0"
   },
   "main": "Parse-Dashboard/app.js",
   "jest": {
