@@ -8,7 +8,7 @@
 // Command line tool for npm start
 'use strict'
 const path = require('path');
-const jsonFile = require('json-file-plus');
+const fs = require('fs');
 const express = require('express');
 const parseDashboard = require('./app');
 const CLIHelper = require('./CLIHelper.js');
@@ -126,74 +126,72 @@ if (!program.config && !process.env.PARSE_DASHBOARD_CONFIG) {
   }
 }
 
-let p = null;
+let config = null;
 let configFilePath = null;
 if (configFile) {
-  p = jsonFile(configFile);
-  configFilePath = path.dirname(configFile);
+  try {
+    config = {
+      data: JSON.parse(fs.readFileSync(configFile, 'utf8'))
+    };
+    configFilePath = path.dirname(configFile);
+  } catch (error) {
+    if (error instanceof SyntaxError) {
+      console.log('Your config file contains invalid JSON. Exiting.');
+      process.exit(1);
+    } else if (error.code === 'ENOENT') {
+      if (explicitConfigFileProvided) {
+        console.log('Your config file is missing. Exiting.');
+        process.exit(2);
+      } else {
+        console.log('You must provide either a config file or required CLI options (app ID, Master Key, and server URL); not both.');
+        process.exit(3);
+      }
+    } else {
+      console.log('There was a problem with your config. Exiting.');
+      process.exit(-1);
+    }
+  }
 } else if (configFromCLI) {
-  p = Promise.resolve(configFromCLI);
+  config = configFromCLI;
 } else {
   //Failed to load default config file.
   console.log('You must provide either a config file or an app ID, Master Key, and server URL. See parse-dashboard --help for details.');
   process.exit(4);
 }
-p.then(config => {
-  config.data.apps.forEach(app => {
-    if (!app.appName) {
-      app.appName = app.appId;
-    }
-  });
 
-  if (config.data.iconsFolder && configFilePath) {
-    config.data.iconsFolder = path.join(configFilePath, config.data.iconsFolder);
+config.data.apps.forEach(app => {
+  if (!app.appName) {
+    app.appName = app.appId;
   }
-
-  const app = express();
-
-  if (allowInsecureHTTP || trustProxy || dev) app.enable('trust proxy');
-
-  config.data.trustProxy = trustProxy;
-  let dashboardOptions = { allowInsecureHTTP, cookieSessionSecret, dev };
-  app.use(mountPath, parseDashboard(config.data, dashboardOptions));
-  let server;
-  if(!configSSLKey || !configSSLCert){
-    // Start the server.
-    server = app.listen(port, host, function () {
-      console.log(`The dashboard is now available at http://${server.address().address}:${server.address().port}${mountPath}`);
-    });
-  } else {
-    // Start the server using SSL.
-    var fs = require('fs');
-    var privateKey = fs.readFileSync(configSSLKey);
-    var certificate = fs.readFileSync(configSSLCert);
-
-    server = require('https').createServer({
-      key: privateKey,
-      cert: certificate
-    }, app).listen(port, host, function () {
-      console.log(`The dashboard is now available at https://${server.address().address}:${server.address().port}${mountPath}`);
-    });
-  }
-  handleSIGs(server);
-}, error => {
-  if (error instanceof SyntaxError) {
-    console.log('Your config file contains invalid JSON. Exiting.');
-    process.exit(1);
-  } else if (error.code === 'ENOENT') {
-    if (explicitConfigFileProvided) {
-      console.log('Your config file is missing. Exiting.');
-      process.exit(2);
-    } else {
-      console.log('You must provide either a config file or required CLI options (app ID, Master Key, and server URL); not both.');
-      process.exit(3);
-    }
-  } else {
-    console.log('There was a problem with your config. Exiting.');
-    process.exit(-1);
-  }
-})
-.catch(error => {
-  console.log('There was a problem loading the dashboard. Exiting.', error);
-  process.exit(-1);
 });
+
+if (config.data.iconsFolder && configFilePath) {
+  config.data.iconsFolder = path.join(configFilePath, config.data.iconsFolder);
+}
+
+const app = express();
+
+if (allowInsecureHTTP || trustProxy || dev) app.enable('trust proxy');
+
+config.data.trustProxy = trustProxy;
+let dashboardOptions = { allowInsecureHTTP, cookieSessionSecret, dev };
+app.use(mountPath, parseDashboard(config.data, dashboardOptions));
+let server;
+if(!configSSLKey || !configSSLCert){
+  // Start the server.
+  server = app.listen(port, host, function () {
+    console.log(`The dashboard is now available at http://${server.address().address}:${server.address().port}${mountPath}`);
+  });
+} else {
+  // Start the server using SSL.
+  var privateKey = fs.readFileSync(configSSLKey);
+  var certificate = fs.readFileSync(configSSLCert);
+
+  server = require('https').createServer({
+    key: privateKey,
+    cert: certificate
+  }, app).listen(port, host, function () {
+    console.log(`The dashboard is now available at https://${server.address().address}:${server.address().port}${mountPath}`);
+  });
+}
+handleSIGs(server);
