@@ -36,7 +36,6 @@ import Toggle                  from 'components/Toggle/Toggle.react';
 import Toolbar                 from 'components/Toolbar/Toolbar.react';
 import { Directions }          from 'lib/Constants';
 import { extractExpiration, extractPushTime } from 'lib/extractTime';
-import * as queryString        from 'query-string';
 
 const PARSE_SERVER_SUPPORTS_AB_TESTING = false;
 
@@ -148,11 +147,11 @@ class PushNew extends DashboardView {
   componentWillMount() {
     this.props.schema.dispatch(SchemaStore.ActionTypes.FETCH);
     let options = { xhrKey: XHR_KEY };
-    const query = queryString.parse(this.props.location.search);
-    if (query.audienceId) {
+    const query = new URLSearchParams(this.props.location.search);
+    if (query.has('audienceId')) {
       options.limit = PushConstants.SHOW_MORE_LIMIT;
       options.min = PushConstants.INITIAL_PAGE_SIZE;
-      this.setState({ initialAudienceId: query.audienceId });
+      this.setState({ initialAudienceId: query.get('audienceId') });
     }
     this.props.pushaudiences.dispatch(PushAudiencesStore.ActionTypes.FETCH,
       options).then(() => {
